pymoose/requirements-dev.txt
absl-py==0.11.0
black==19.10b0
flake8-black==0.1.1
flake8-isort==4.0.0
flake8==3.8.3
graphviz==0.15
grpcio==1.33.1
grpcio-tools==1.33.1
hypothesis==5.47.0
isort==5.4.2
msgpack==1.0.2
numpy==1.19.0
opentelemetry-api==0.15b0
opentelemetry-launcher==0.15b0
pip~=21.2
pysodium==0.7.5
pytest==6.0.1
pytest-asyncio==0.14.0
pyyaml==5.4
requests==2.24.0
responses==0.12.1
setuptools-rust~=0.11.5
